pyproject.toml
[project]
name = "mock-login"
dynamic = ["version"]
description = "A mock service for local development purposes to substitute the OneAll social login service"
authors = [{ name = "Swindon Makerspace", email = "info@swindon-makerspace.org" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "aiohttp (>=3.11.12,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
requires-poetry = ">=2.0"
version = "0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
